+++ GenericEventRunner/ForSetup/Internal/RegisterIfNotThere.cs
@@ -1,0 +1,14 @@
+﻿// Copyright (c) 2020 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
+// Licensed under MIT license. See License.txt in the project root for license information.
+
+using System.Collections.Generic;
+
+namespace GenericEventRunner.ForSetup.Internal
+{
+    public class RegisterIfNotThere
+    {
+        public List<string> DebugLogs = new List<string>();
+
+
+    }
+}